requirements.txt
anthropic==0.7.8
openai==1.5.0
python-dotenv==1.0.0
flask==2.3.3
pydantic==2.5.2
colorlog==6.7.0
tenacity==8.2.3 